src/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
djangorestframework==3.14.0
drf-yasg==1.21.7
flake8==6.1.0
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
isort==5.12.0
mccabe==0.7.0
multidict==6.0.4
packaging==23.2
psycopg2==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
yarl==1.9.2
